+++ lib/polars/series.rb
@@ -2571,8 +2571,6 @@ module Polars
             Float32 => Numo::SFloat,
             Float64 => Numo::DFloat
           }.fetch(dtype.class).cast(to_a)
-        elsif is_boolean
-          Numo::Bit.cast(to_a)
         else
           _s.to_numo
         end